+++ resources/lib/commands/romcollection_roms_commands.py
@@ -58,12 +58,12 @@ def cmd_manage_roms(args):
     options = collections.OrderedDict()
     options['SET_ROMS_DEFAULT_ARTWORK'] = kodi.translate(42044)
     
-    if romcollection.has_scanners():
-        options['SCAN_ROMS'] = kodi.translate(42046)
-        options['REMOVE_DEAD_ROMS'] = kodi.translate(42047)
-        options['EDIT_ROMCOLLECTION_SCANNERS'] = kodi.translate(42048)
-    else:
-        options['ADD_SCANNER'] = kodi.translate(42049)
+    # if romcollection.has_scanners():
+    #     options['SCAN_ROMS'] = kodi.translate(42046)
+    #     options['REMOVE_DEAD_ROMS'] = kodi.translate(42047)
+    #     options['EDIT_ROMCOLLECTION_SCANNERS'] = kodi.translate(42048)
+    # else:
+    #     options['ADD_SCANNER'] = kodi.translate(42049)
     
     options['IMPORT_ROMS'] = kodi.translate(42050)
     if has_roms:
